+++ src/getStepStartStates.ts
@@ -70,7 +70,10 @@ import {
   STEP_BUILD_BIG_QUERY_DATASET_KMS_RELATIONSHIPS,
 } from './steps/big-query';
 import { STEP_SQL_ADMIN_INSTANCES } from './steps/sql-admin';
-import { STEP_DNS_MANAGED_ZONES } from './steps/dns/constants';
+import {
+  STEP_DNS_MANAGED_ZONES,
+  STEP_DNS_POLICIES,
+} from './steps/dns/constants';
 import { STEP_CONTAINER_CLUSTERS } from './steps/containers';
 import {
   STEP_LOGGING_METRICS,
@@ -391,6 +394,7 @@ export default async function getStepStartStates(
       ServiceUsageName.SQL_ADMIN,
     ),
     [STEP_DNS_MANAGED_ZONES]: createStepStartState(ServiceUsageName.DNS),
+    [STEP_DNS_POLICIES]: createStepStartState(ServiceUsageName.DNS),
     [STEP_CONTAINER_CLUSTERS]: createStepStartState(ServiceUsageName.CONTAINER),
     [STEP_LOGGING_PROJECT_SINKS]: createStepStartState(
       ServiceUsageName.LOGGING,