+++ CHANGELOG.md
@@ -8,6 +8,31 @@ and this project adheres to
 
 ## [Unreleased]
 
+## 0.51.4 - 2021-09-27
+
+### Added
+
+- Initial managed JupiterOne questions moved into this project
+
+### Fixed
+
+- `Internet` **ALLOWS** `google_compute_firewall` relationship for
+  `0.0.0.0/0`/`::/0` source CIDR blocks restored by adding `_type` to the target
+  filter keys.
+
+## 0.50.0 - 2021-09-15
+
+### Added
+
+- New properties added to resources:
+
+  | Entity                           | Properties    |
+  | -------------------------------- | ------------- |
+  | `google_bigquery_table`          | `kmsKeyName`  |
+  | `google_sql_sql_server_instance` | `userOptions` |
+
+## 0.49.0 - 2021-09-14
+
 ### Added
 
 - Create relationships for every member of `google_iam_binding`s.
@@ -72,6 +97,18 @@ and this project adheres to
 ## 0.48.0 - 2021-08-27
 
 ### Changed
+
+- Added support for ingesting the following **new** resources:
+
+  | Service    | Resource / Entity   |
+  | ---------- | ------------------- |
+  | DNS Policy | `google_dns_policy` |
+
+- Added support for ingesting the following **new** relationships:
+
+  | Source                   | class   | Target              |
+  | ------------------------ | ------- | ------------------- |
+  | `google_compute_network` | **HAS** | `google_dns_policy` |
 
 - Relationships from `google_cloud_organization`s and `google_cloud_folder`s to
   `google_cloud_project`s will also be made for deleted projects as well.